Refactor ProcessService to improve command handling and add stdout/stderr accessors

+++ src/process/service.ts
@@ -1,19 +1,8 @@
 import { Injectable } from '@danet/core';
-import {ProcessInfo, ProcessRequest} from "./interface.ts";
-
+import {ProcessInfo} from "./interface.ts";
 
 /**
- * Lethean ProcessManager handles all aspects of running external binaries
- * you need to provide the correct binary for the OS, all other host differences are handled for you
- * @example
- * ProcessManager.run(path.join(homeDir, 'Lethean', 'cli', exeFile),args,
- * {
- * 		key: exeFile,
- * 		stdErr: (stdErr: unknown) => console.log(stdErr),
- * 		stdIn: (stdIn: unknown) => console.log(stdIn),
- * 		stdOut: (stdOut: unknown) => console.log(stdOut)
- * 	} as ProcessInterface
- * 	);
+ * Process Service
  */
 @Injectable()
 export class ProcessService {
@@ -24,6 +13,8 @@ export class ProcessService {
     private command: Deno.Command | undefined;
     private args!: string[];
     private handle!: Deno.ChildProcess;
+    private _stdout!: ReadableStream<Uint8Array>;
+    private _stderr!: ReadableStream<Uint8Array>;
     private _info: ProcessInfo  = {
         command: '',
         args: [],
@@ -36,6 +27,11 @@ export class ProcessService {
         return this._info;
     }
 
+    /**
+     * Create a Deno.Command object
+     * @param command The full path to the binary
+     * @param args The arguments to pass to the binary
+     */
     add(command: string, args: string[] = []){
         if(command.length === 0){
             return false;
@@ -48,6 +44,11 @@ export class ProcessService {
         return this.command = new Deno.Command(command, {args: this.args})
     }
 
+    /**
+     * Run a command with arguments & wait for its return value
+     * @param command The full path to the binary
+     * @param args The arguments to pass to the binary
+     */
     async run(command: string, args: string[] = []){
         this._info.command = command;
         if(args.length > 0){
@@ -61,71 +62,32 @@ export class ProcessService {
         return { code, stdout, stderr }
     }
 
-    start(command: string, args?: string[]){
-        //return this.handle = this.command.spawn()
+
+    start(command: string, args: string[]){
+        this._info.command = command;
+        if(args.length > 0){
+            this._info.args = args;
+        }
+        this._info.time_added = Date.now();
+        this.command = new Deno.Command(command, {args: args})
+        this._info.time_started = Date.now();
+        const { stdout, stderr } = this.command.spawn()
+        this._info.time_stopped = Date.now();
+        this._stdout = stdout;
+        this._stderr = stderr;
+        return { stdout, stderr }
     }
 
     stop(){
-        //this.has(command, args)
+        if(!this.handle){
+            return false;
+        }
         return this.handle.kill()
     }
-
-
-
-
-    /**
-     * Creates a src record and then starts it...
-     * to be expanded as you can do ProcessManager.startProcess('letheand.exe')
-     * so starting right away is optional
-     */
-    // startmm(command: string, args: any, options?: ProcessRequest) {
-    //     // if (!args) {
-    //     //     console.log("No arguments passed to ProcessManager");
-    //     //     return;
-    //     // }
-    //     //
-    //     // const cmdArgs = [command];
-    //     //
-    //     // for (const arg in args) {
-    //     //     if (arg === "igd") {
-    //     //         continue;
-    //     //     }
-    //     //     cmdArgs.push(args[arg]);
-    //     //     // cmdArgs.push(
-    //     //     //     "--" + arg.replace(/([A-Z])/g, (x) => "-" + x.toLowerCase()) +
-    //     //     //     (args[arg].length > 1 ? `=${args[arg]}` : ""),
-    //     //     // );
-    //     // }
-    //
-    //     // if(!options){
-    //     //     // options = {
-    //     //     //     key: 'test',
-    //     //     //     command: cmdArgs,
-    //     //     //     stdErr: (stdErr: unknown) => console.log(stdErr),
-    //     //     //     stdIn: (stdIn: unknown) => console.log(stdIn),
-    //     //     //     stdOut: (stdOut: unknown) => console.log(stdOut)
-    //     //     // } as ProcessRequest;
-    //     // }
-    //
-    //     // if (options.key && ProcessService.process[options.key]) {
-    //     //     return this.getProcess(options.key);
-    //     // }
-    //
-    //     // if (ProcessService.debug) {
-    //     //     console.log("Arguments passed to ProcessManager:", args);
-    //     // }
-    //     //
-    //     //
-    //     // if (ProcessService.debug) {
-    //     //     console.log(
-    //     //         "ProcessManager processed arguments to these:",
-    //     //         cmdArgs,
-    //     //     );
-    //     // }
-    //
-    //     // return this.addProcess(options).run();
-    // }
-
-
-
+    stdout(){
+        return this._stdout;
+    }
+    stderr(){
+        return this._stderr;
+    }
 }
